delete workers and new theme

+++ lib/src/widgets/editarProyecto.dart
@@ -157,19 +157,47 @@ class _DetallesYAltaProyectoPageState extends State<DetallesYAltaProyectoPage> {
             ),
             const SizedBox(height: 16.0),
             Container(
-                width: double.infinity,
-                child: IconButton.filled(
-                  icon: Icon(Icons.delete_forever),
-                  onPressed: () async {
-                    try {
-                      await deleteProyecto(widget.proyecto?['id']);
-                      print('Proyecto eliminado');
-                      Navigator.pop(context);
-                    } catch (error) {
-                      print('Error al eliminar el proyecto: $error');
-                    }
-                  },
-                )),
+              width: double.infinity,
+              child: IconButton.filled(
+                icon: Icon(Icons.delete_forever),
+                onPressed: () async {
+                  // Mostrar un diálogo de confirmación antes de eliminar
+                  showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return AlertDialog(
+                        title: Text('Eliminar Proyecto'),
+                        content:
+                            Text('¿Seguro que quieres eliminar este proyecto?'),
+                        actions: <Widget>[
+                          TextButton(
+                            onPressed: () {
+                              Navigator.of(context).pop();
+                            },
+                            child: Text('Cancelar'),
+                          ),
+                          TextButton(
+                            onPressed: () async {
+                              try {
+                                Navigator.of(context)
+                                    .pop(); // Cerrar el diálogo antes de la eliminación
+                                await deleteProyecto(widget.proyecto?['id']);
+                                print('Proyecto eliminado');
+                                Navigator.pop(context);
+                                // Puedes agregar más lógica aquí si es necesario
+                              } catch (error) {
+                                print('Error al eliminar el proyecto: $error');
+                              }
+                            },
+                            child: Text('Confirmar'),
+                          ),
+                        ],
+                      );
+                    },
+                  );
+                },
+              ),
+            ),
           ],
         ),
       ),
